requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.2
django-celery-beat==2.5.0
django-cors-headers==4.1.0
django-crontab==0.7.1
django-filter==23.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.6
numpy==1.26.4
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.6
PyJWT==2.8.0
pyscopg2==66.0.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0

